+++ flutter_theoplayer_sdk/flutter_theoplayer_sdk_android/CHANGELOG.md
@@ -1,3 +1,17 @@
+## 9.11.0
+
+* Updated THEOplayer to 9.11.0.
+
+## 9.7.1
+
+* Added `TypedSource.type` property.
+* Added `WebConfig.libraryLocation` to specify a different path for THEOplayer WEB SDK through `THEOplayerConfig`.
+
+## 9.7.0
+
+* Updated THEOplayer to 9.7.0.
+* Added `TheoLiveConfiguration.discoveryUrl` to support custom discovery URLs for OptiView live streams (formerly known as THEOlive). 
+
 ## 9.3.1
 
 * Updated THEOplayer to 9.3.1.